Change 2
The Icon and drop down menu icon are not appearing. I have gone back to
the First Upload and they do not appear. The problem may be occuring
because of Notepad++ or the way I am launching the site (Google Chrome).
Changes: mobile version now does not need to have a side to side scroll
bar anymore.